Fix #1911 : Create a separate package for ApiExplorer
