+++ samples/MvcSample.Web/Views/ApiExplorer/All.cshtml
@@ -1,4 +1,4 @@
-﻿@using Microsoft.AspNet.Mvc.Description
+﻿@using Microsoft.AspNet.Mvc.ApiExplorer
 @model IReadOnlyList<ApiDescriptionGroup>
 
 @section header
