+++ samples/MvcSample.Web/Views/ApiExplorer/_ApiDescription.cshtml
@@ -1,4 +1,4 @@
-﻿@using Microsoft.AspNet.Mvc.Description
+﻿@using Microsoft.AspNet.Mvc.ApiExplorer
 @model ApiDescription
 
 <div class="api-description">
