+++ src/Microsoft.AspNet.Mvc.ApiExplorer/ApiDescriptionExtensions.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Framework.Internal;
 
-namespace Microsoft.AspNet.Mvc.Description
+namespace Microsoft.AspNet.Mvc.ApiExplorer
 {
     /// <summary>
     /// Extension methods for <see cref="ApiDescription"/>.
