+++ src/Microsoft.AspNet.Mvc.ApiExplorer/ApiDescriptionGroup.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 
-namespace Microsoft.AspNet.Mvc.Description
+namespace Microsoft.AspNet.Mvc.ApiExplorer
 {
     /// <summary>
     /// Represents a group of related apis.
