+++ src/Microsoft.AspNet.Mvc.ApiExplorer/ApiDescriptionGroupCollection.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using Microsoft.Framework.Internal;
 
-namespace Microsoft.AspNet.Mvc.Description
+namespace Microsoft.AspNet.Mvc.ApiExplorer
 {
     /// <summary>
     /// A cached collection of <see cref="ApiDescriptionGroup" />.
