+++ src/Microsoft.AspNet.Mvc.ApiExplorer/ApiExplorerSettingsAttribute.cs
@@ -2,13 +2,13 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.AspNet.Mvc.Description;
+using Microsoft.AspNet.Mvc.ApiExplorer;
 
 namespace Microsoft.AspNet.Mvc
 {
     /// <summary>
-    /// Controls the visibility and group name for an <see cref="ApiDescription"/> of the associated
-    /// controller class or action method.
+    /// Controls the visibility and group name for an <see cref="ApiDescription"/>
+    /// of the associated controller class or action method.
     /// </summary>
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = false, Inherited = true)]
     public class ApiExplorerSettingsAttribute :
