+++ src/Microsoft.AspNet.Mvc.ApiExplorer/ApiParameterRouteInfo.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using Microsoft.AspNet.Routing;
 
-namespace Microsoft.AspNet.Mvc.Description
+namespace Microsoft.AspNet.Mvc.ApiExplorer
 {
     /// <summary>
     /// A metadata description of routing information for an <see cref="ApiParameterDescription"/>.
@@ -31,9 +31,9 @@ namespace Microsoft.AspNet.Mvc.Description
         /// <remarks>
         /// An optional parameter is considered optional by the routing system. This does not imply
         /// that the parameter is considered optional by the action.
-        /// 
+        ///
         /// If the parameter uses <see cref="ModelBinding.BindingSource.ModelBinding"/> for the value of
-        /// <see cref="ApiParameterDescription.Source"/> then the value may also come from the 
+        /// <see cref="ApiParameterDescription.Source"/> then the value may also come from the
         /// URL query string or form data.
         /// </remarks>
         public bool IsOptional { get; set; }
