+++ src/Microsoft.AspNet.Mvc.ApiExplorer/IApiDescriptionProvider.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Framework.Internal;
 
-namespace Microsoft.AspNet.Mvc.Description
+namespace Microsoft.AspNet.Mvc.ApiExplorer
 {
     public interface IApiDescriptionProvider
     {
