+++ src/Microsoft.AspNet.Mvc.Core/DefaultAssemblyProvider.cs
@@ -30,7 +30,8 @@ namespace Microsoft.AspNet.Mvc
         protected virtual HashSet<string> ReferenceAssemblies { get; } = new HashSet<string>(StringComparer.Ordinal)
         {
             "Microsoft.AspNet.Mvc",
-            "Microsoft.AspNet.Mvc.Abstractions",
+             "Microsoft.AspNet.Mvc.Abstractions",
+            "Microsoft.AspNet.Mvc.ApiExplorer",
             "Microsoft.AspNet.Mvc.Core",
             "Microsoft.AspNet.Mvc.Razor",
             "Microsoft.AspNet.Mvc.Razor.Host",
