+++ src/Microsoft.AspNet.Mvc/project.json
@@ -7,6 +7,7 @@
     "dependencies": {
         "Microsoft.AspNet.Authorization": "1.0.0-*",
         "Microsoft.AspNet.Cors": "1.0.0-*",
+        "Microsoft.AspNet.Mvc.ApiExplorer": "6.0.0-*",
         "Microsoft.AspNet.Mvc.Common": { "version": "6.0.0-*", "type": "build" },
         "Microsoft.AspNet.Mvc.Razor": "6.0.0-*",
         "Microsoft.Framework.Caching.Memory": "1.0.0-*",
